# internal/swagger/swagger.yaml
basePath: /
definitions:
...
  dto.AddRoleDTO:
    properties:
      menu_ids:
        items:
          type: integer
        type: array
      name:
        maxLength: 50
        minLength: 1
        type: string
      remark:
        maxLength: 200
        type: string
      status:
        description: 0-禁用 1-启用
        enum:
        - 0
        - 1
        type: integer
    required:
    - name
    type: object
  dto.AssignMenusDTO:
    properties:
      menu_ids:
        items:
          type: integer
        minItems: 1
        type: array
      role_id:
        type: integer
    required:
    - menu_ids
    - role_id
    type: object
...
  dto.AssignRolesDTO:
    properties:
      role_ids:
        items:
          type: integer
        minItems: 1
        type: array
      user_id:
        type: integer
    required:
    - role_ids
    - user_id
    type: object
...
  dto.DeletePermissionDTO:
    properties:
      menu_id:
        type: integer
      perm_key:
        type: string
    required:
    - menu_id
    - perm_key
    type: object
...
  dto.EditRoleDTO:
    properties:
      id:
        type: integer
      menu_ids:
        items:
          type: integer
        type: array
      name:
        maxLength: 50
        minLength: 1
        type: string
      remark:
        maxLength: 200
        type: string
      status:
        description: 0-禁用 1-启用
        enum:
        - 0
        - 1
        type: integer
    required:
    - id
    type: object
  dto.GetRoleDTO:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  dto.GetRolePermissionsDTO:
    properties:
      roleID:
        type: integer
    required:
    - roleID
    type: object
  dto.GetUserRolesDTO:
    properties:
      userID:
        type: integer
    required:
    - userID
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      request_id:
        type: string
    type: object
info:
  contact: {}
  description: webgos 企业资源计划系统 API 文档
  title: webgos API
  version: "1.0"
paths:
...
  /api/rbac/assign_menus:
    post:
      consumes:
      - application/json
      description: 给角色分配菜单，角色通过菜单间接获得权限
      parameters:
      - description: 菜单分配信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AssignMenusDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 分配菜单给角色
      tags:
      - 角色权限
  /api/rbac/assign_roles:
    post:
      consumes:
      - application/json
      description: 给用户分配角色
      parameters:
      - description: 角色分配信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AssignRolesDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 分配角色给用户
      tags:
      - 角色权限
  /api/rbac/edit_role:
    post:
      consumes:
      - application/json
      description: 更新角色信息
      parameters:
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.EditRoleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 编辑角色
      tags:
      - 角色权限
  /api/rbac/permission:
    delete:
      consumes:
      - application/json
      description: 按 menu_id + perm_key 解绑单个菜单权限键
      parameters:
      - description: 解绑信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.DeletePermissionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 解绑菜单权限
      tags:
      - 角色权限
  /api/rbac/permissions:
    get:
      consumes:
      - application/json
      description: 获取所有权限列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取权限列表
      tags:
      - 角色权限
  /api/rbac/role:
    post:
      consumes:
      - application/json
      description: 创建新角色
      parameters:
      - description: 角色信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddRoleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - 角色权限
  /api/rbac/role/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取角色详情
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.GetRoleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取角色详情
      tags:
      - 角色权限
  /api/rbac/role_permissions/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定角色的权限列表
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.GetRolePermissionsDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取角色的权限列表
      tags:
      - 角色权限
  /api/rbac/roles:
    get:
      consumes:
      - application/json
      description: 获取所有角色列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取角色列表
      tags:
      - 角色权限
  /api/rbac/user_roles/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定用户的角色列表
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.GetUserRolesDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户的角色列表
      tags:
      - 角色权限
...
securityDefinitions:
  BearerAuth:
    description: JWT认证方式，值为"Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"